Move GroupChecker into group.go
